Purged obsolete data-c++ library because it caused crashes on amd64 architecture
git-svn-id: file:///scratch/steve/svn/rapidSTORM-dev/rapidSTORM/trunk@3274 6dd997e4-92ef-4477-b1f0-deb9b3b8d6d5

+++ test-plugin/Manager.cpp
@@ -247,7 +247,7 @@ bool Manager::Source::get_and_handle_change() {
     }
 
     for ( size_t j = 0; j < c->changed_keys.size(); ++j)
-        for ( int i = 0; i < c->changed_keys[j].size(); i++ )
+        for ( size_t i = 0; i < c->changed_keys[j].size(); i++ )
         {
             dStorm::Display::KeyChange kc = c->changed_keys[j][i];
             if ( i == 31 )
